refactor: tests

+++ test/test.sh
@@ -1,17 +1,10 @@
 #!/bin/sh
 
-cd test
-
-if [ -f freeze-test ]; then
-  rm freeze-test
-fi
-
+[ -d test ] && cd test
+[ -f freeze-test ] && rm freeze-test
 go build -o freeze-test ..
-
-if [ ! -f freeze-test ]; then
-  echo "Failed to build freeze-test"
-  exit 1
-fi
+[ ! -f freeze-test ] && echo "Failed to build freeze-test" && exit 1
+[ ! -d output ] && mkdir output/
 
 if [ ! -d golden ]; then
   echo "No golden files, generating..."
@@ -27,33 +20,20 @@ if [ ! -d golden ]; then
   exit 0
 fi
 
-if [ ! -d output ]; then
-  mkdir output/
-fi
 
 FAILURES=0
 for f in configurations/*.json; do
   filename=$(basename -- "$f")
-
   if [ ! -f golden/"${filename%.*}".svg ]; then
     echo "Generating golden file for $filename, verify and commit."
     ./freeze-test --config $f --output golden/"${filename%.*}".svg $f
     continue
   fi
-
   ./freeze-test --config $f --output output/"${filename%.*}".svg $f
-
   diff --color output/"${filename%.*}".svg golden/"${filename%.*}".svg
-  if [ $? -ne 0 ]; then
-    echo "=== Test failed for $filename ==="
-    FAILURES=$((FAILURES + 1))
-  fi
+  [ $? -ne 0 ] && echo "=== Test failed for $filename ===" &&  FAILURES=$((FAILURES + 1))
 done
 
-if [ $FAILURES -ne 0 ]; then
-  echo "$FAILURES tests failed"
-  rm freeze-test
-  exit 1
-fi
-
 rm freeze-test
+
+[ $FAILURES -ne 0 ] && echo "$FAILURES tests failed" && exit 1
